Fix issue JAX-RS JSON serialization not working in shaded jar
When running the shaded JAR, I was getting this error: ERROR  MessageBodyWriter not found for media type=application/json, type=class io.github.milobotdev.milobot.api.LoginReturnData, genericType=class io.github.milobotdev.milobot.api.LoginReturnData.

The jersey-media-json-jackson dependency should fix this: https://stackoverflow.com/questions/26207252/messagebodywriter-not-found-for-media-type-application-json

However, it only works when I run it in IntelliJ IDEA, I still get the error when I run the shaded jar.

So, according to this Stack Overflow answer: https://stackoverflow.com/questions/44552733/jersey-maven-messagebodywriter-not-found-for-media-type-application-json/52291171 jersey-media-json-jackson has to be the first dependency.

And... this fixed the issue

It's a weird fix, the second answer has some explanation to why this fixes it.